Initial changes for Authorizer interface

+++ hadoop-tools/hadoop-azure/src/test/java/org/apache/hadoop/fs/azurebfs/AbstractAbfsIntegrationTest.java
@@ -24,6 +24,11 @@ import java.util.Hashtable;
 import java.util.UUID;
 import java.util.concurrent.Callable;
 
+import org.apache.hadoop.fs.azurebfs.constants.AbfsHttpConstants;
+import org.apache.hadoop.fs.azurebfs.contracts.exceptions.InvalidConfigurationValueException;
+import org.apache.hadoop.fs.azurebfs.contracts.exceptions.InvalidUriAuthorityException;
+import org.apache.hadoop.fs.azurebfs.contracts.exceptions.InvalidUriException;
+import org.apache.hadoop.fs.azurebfs.extensions.AbfsAuthorizer;
 import org.junit.After;
 import org.junit.Before;
 import org.slf4j.Logger;
@@ -73,6 +78,7 @@ public abstract class AbstractAbfsIntegrationTest extends
   private String accountName;
   private String testUrl;
   private AuthType authType;
+  private AbfsAuthorizer authorizer;
 
   protected AbstractAbfsIntegrationTest() throws Exception {
     fileSystemName = TEST_CONTAINER_PREFIX + UUID.randomUUID().toString();
@@ -128,6 +134,46 @@ public abstract class AbstractAbfsIntegrationTest extends
   }
 
 
+  public void loadAuthorizer() throws Exception {
+    this.authorizer = abfsConfig.getAbfsAuthorizer();
+
+    // if authorizerClass is configured, auto-creation of filesystem is disabled
+    if (authorizer == null) {
+      throw new InvalidConfigurationValueException(
+          "loadAuthorizer failed as " + "authorizer class is not configured");
+    }
+
+    // if authorizerClass is configured, auto-creation of filesystem is disabled
+    abfsConfig.setBoolean(AZURE_CREATE_REMOTE_FILESYSTEM_DURING_INITIALIZATION,
+        false);
+
+    // AbstractAbfsIntegrationTest always uses a new instance of FileSystem,
+    // need to disable that and force filesystem provided in test configs.
+    String[] authorityParts = authorityParts(
+        new java.net.URI(rawConfig.get(FS_AZURE_CONTRACT_TEST_URI)));
+    this.fileSystemName = authorityParts[0];
+
+    // Reset URL with configured filesystem
+    final String abfsUrl =
+        this.getFileSystemName() + "@" + this.getAccountName();
+    URI defaultUri = null;
+
+    try {
+      defaultUri = new URI(abfsScheme, abfsUrl, null, null, null);
+    } catch (Exception ex) {
+      throw new AssertionError(ex);
+    }
+
+    this.testUrl = defaultUri.toString();
+    abfsConfig.set(CommonConfigurationKeysPublic.FS_DEFAULT_NAME_KEY,
+        defaultUri.toString());
+
+    // update rawConfig which will be used to fetch the filesystem instance
+    rawConfig = abfsConfig.getRawConfiguration();
+    rawConfig.set(FS_AZURE_ABFS_AUTHORIZER, abfsConfig.getAbfsAuthorizerClass());
+  }
+
+
   @Before
   public void setup() throws Exception {
     //Create filesystem first to make sure getWasbFileSystem() can return an existing filesystem.
@@ -167,19 +213,22 @@ public abstract class AbstractAbfsIntegrationTest extends
         return;
       }
 
-      final AzureBlobFileSystemStore abfsStore = abfs.getAbfsStore();
-      abfsStore.deleteFilesystem();
+      // If authorizer is present, configured FileSystem would have been used,
+      // instead of unique. Skip Delete.
+      if (authorizer == null) {
+        final AzureBlobFileSystemStore abfsStore = abfs.getAbfsStore();
+        abfsStore.deleteFilesystem();
 
-      AbfsRestOperationException ex = intercept(
-              AbfsRestOperationException.class,
-              new Callable<Hashtable<String, String>>() {
-                @Override
-                public Hashtable<String, String> call() throws Exception {
-                  return abfsStore.getFilesystemProperties();
-                }
-              });
-      if (FILE_SYSTEM_NOT_FOUND.getStatusCode() != ex.getStatusCode()) {
-        LOG.warn("Deleted test filesystem may still exist: {}", abfs, ex);
+        AbfsRestOperationException ex = intercept(AbfsRestOperationException.class,
+            new Callable<Hashtable<String, String>>() {
+              @Override
+              public Hashtable<String, String> call() throws Exception {
+                return abfsStore.getFilesystemProperties();
+              }
+            });
+        if (FILE_SYSTEM_NOT_FOUND.getStatusCode() != ex.getStatusCode()) {
+          LOG.warn("Deleted test filesystem may still exist: {}", abfs, ex);
+        }
       }
     } catch (Exception e) {
       LOG.warn("During cleanup: {}", e, e);
@@ -350,5 +399,31 @@ public abstract class AbstractAbfsIntegrationTest extends
   protected AbfsDelegationTokenManager getDelegationTokenManager()
       throws IOException {
     return getFileSystem().getDelegationTokenManager();
+  }
+
+  private String[] authorityParts(URI uri)
+      throws InvalidUriAuthorityException, InvalidUriException {
+    final String authority = uri.getRawAuthority();
+    if (null == authority) {
+      throw new InvalidUriAuthorityException(uri.toString());
+    }
+
+    if (!authority.contains(
+        AbfsHttpConstants.AZURE_DISTRIBUTED_FILE_SYSTEM_AUTHORITY_DELIMITER)) {
+      throw new InvalidUriAuthorityException(uri.toString());
+    }
+
+    final String[] authorityParts = authority.split(
+        AbfsHttpConstants.AZURE_DISTRIBUTED_FILE_SYSTEM_AUTHORITY_DELIMITER, 2);
+
+    if (authorityParts.length < 2
+        || authorityParts[0] != null && authorityParts[0].isEmpty()) {
+      final String errMsg = String.format(
+          "'%s' has a malformed authority, expected container name. "
+              + "Authority takes the form " + FileSystemUriSchemes.ABFS_SCHEME
+              + "://[<container name>@]<account name>", uri.toString());
+      throw new InvalidUriException(errMsg);
+    }
+    return authorityParts;
   }
 }
